requirements.txt
Django
django-dotenv
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-import-export==2.7.1